pyproject.toml
[tool.poetry]
name = "locust-review"
version = "0.1.0"
description = "An overview of using Locust for load testing notifications"
authors = ["cris-oddball <cris.stoddard@oddball.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
locust = "^2.31.0"
pyJWT = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
